requirements.txt
Flask==2.3.3
scikit-learn==1.3.1
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
BeautifulSoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.1.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
regex==2023.6.3
Werkzeug>=2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
email-validator==2.0.0
urllib3==2.0.4
python-magic==0.4.27